Refactor P&ID function (#369)
* Updated to adhere to Functions specifications.

* Split into several files

* Refactored a lot of functionality into smaller functions

* General code modernization and maintenance
