+++ .pre-commit-config.yaml
@@ -12,14 +12,12 @@ repos:
           - --select=E,W,F,I,T,RUF,TID,UP
           - --fixable=E,W,F,I,T,RUF,TID,UP
           - --target-version=py39
-          - --exclude=cognite/client/_proto,cognite/client/_proto_legacy
     rev: v0.2.2
 
   - repo: https://github.com/psf/black
     rev: 24.2.0
     hooks:
       - id: black
-        exclude: ^cognite/client/_proto.*$
         args:
           - --line-length=120
           - --target-version=py38
