+++ cognite_toolkit/cognite_modules/examples/cdf_data_pipeline_files_valhall/functions/fu_context_files_oid_fileshare_annotation/constants.py
@@ -1,0 +1,17 @@
+# P&ID original file defaults
+ORG_MIME_TYPE = "application/pdf"
+FILE_ANNOTATED_META_KEY = "FILE_ANNOTATED"
+ANNOTATION_ERROR_MSG = "annotation_created_error"
+
+# Annotation defaults
+ASSET_ANNOTATION_TYPE = "diagrams.AssetLink"
+FILE_ANNOTATION_TYPE = "diagrams.FileLink"
+ANNOTATION_STATUS_APPROVED = "approved"
+ANNOTATION_STATUS_SUGGESTED = "suggested"
+ANNOTATION_RESOURCE_TYPE = "file"
+CREATING_APP = "P&ID contextualization and annotation function"
+CREATING_APPVERSION = "1.0.0"
+
+# Other constants
+ASSET_MAX_LEN_META = 10000
+ISO_8601 = "%Y-%m-%d %H:%M:%S"
