+++ cognite_toolkit/cognite_modules/examples/cdf_data_pipeline_files_valhall/functions/fu_context_files_oid_fileshare_annotation/requirements.txt
@@ -1,4 +1,2 @@
-cognite-extractor-utils>=6.3.2
-cognite-logger>=0.6
-cognite-sdk>=7.13.8
-pyyaml>=6.0.1
+cognite-sdk == 7.*
+pyyaml >= 6.0.1
