+++ cognite_toolkit/cognite_modules/examples/cdf_data_pipeline_files_valhall/functions/functions.yaml
@@ -1,4 +1,4 @@
-# The dire with the function code should have the same name
+# The directory with the function code should have the same name
 # and externalId as the function itself as defined below.
 - name: 'context:files:{{location_name}}:{{source_name}}:annotation'
   externalId: 'fu_context_files_{{location_name}}_{{source_name}}_annotation'
@@ -17,5 +17,3 @@
   functionPath: './handler.py'
   # Data set id for the zip file with the code that is uploaded.
   externalDataSetId: 'ds_files_{{location_name}}'
-
-
