+++ cognite_toolkit/cognite_modules/examples/cdf_data_pipeline_files_valhall/functions/schedules.yaml
@@ -1,9 +1,9 @@
-- name: 'daily-every-30-min'
-  functionExternalId: 'fu_context_files_{{location_name}}_{{source_name}}_annotation'
-  description: "Run every 30 minute"
+- name: every 30 min
+  functionExternalId: fu_context_files_{{location_name}}_{{source_name}}_annotation
+  description: Run every 30 minute
   cronExpression: "0,30 * * * *"
   data:
-    ExtractionPipelineExtId: 'ep_ctx_files_{{location_name}}_{{source_name}}_pandid_annotation'
+    ExtractionPipelineExtId: ep_ctx_files_{{location_name}}_{{source_name}}_pandid_annotation
   authentication:
     # Credentials to use to run the function in this schedule.
     # In this example, we just use the main deploy credentials, so the result is the same, but use a different set of
