+++ tests/tests_unit/test_approval_modules_snapshots/cdf_data_pipeline_files_valhall.yaml
@@ -76,7 +76,7 @@ Function:
   externalId: fu_context_files_oid_fileshare_annotation
   functionPath: ./handler.py
   metadata:
-    cdf-toolkit-function-hash: 9fc4f831d321aa088af7e0ac295ecc308152afcecf84152a02a04400e404b11c
+    cdf-toolkit-function-hash: b3cbb16c2cbbd790512d7485c95b055423140d59ae0bf39ca763ee2b03c50678
     version: 0.0.1
   name: context:files:oid:fileshare:annotation
   owner: Anonymous
